Create unified parser
