+++ components/generator/jvm/imagevector/src/test/kotlin/io/github/composegears/valkyrie/generator/jvm/imagevector/BrokenSvgParsingTest.kt
@@ -3,7 +3,9 @@ package io.github.composegears.valkyrie.generator.jvm.imagevector
 import assertk.assertFailure
 import assertk.assertions.isInstanceOf
 import io.github.composegears.valkyrie.extensions.ResourceUtils.getResourcePath
-import io.github.composegears.valkyrie.parser.svgxml.SvgXmlParser
+import io.github.composegears.valkyrie.parser.unified.ParserType
+import io.github.composegears.valkyrie.parser.unified.SvgXmlParser
+import io.github.composegears.valkyrie.parser.unified.ext.toIOPath
 import java.io.EOFException
 import org.junit.jupiter.api.Test
 
@@ -11,8 +13,10 @@ class BrokenSvgParsingTest {
 
     @Test
     fun `parsing broken svg`() {
-        val icon = getResourcePath("imagevector/broken.svg")
+        val icon = getResourcePath("imagevector/broken.svg").toIOPath()
 
-        assertFailure { SvgXmlParser.toIrImageVector(icon) }.isInstanceOf(EOFException::class)
+        assertFailure {
+            SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
+        }.isInstanceOf(EOFException::class)
     }
 }
