+++ components/generator/jvm/imagevector/src/test/kotlin/io/github/composegears/valkyrie/generator/jvm/imagevector/ImageVectorWithIndentTest.kt
@@ -5,8 +5,10 @@ import assertk.assertions.isEqualTo
 import io.github.composegears.valkyrie.extensions.ResourceUtils.getResourcePath
 import io.github.composegears.valkyrie.generator.jvm.imagevector.common.createConfig
 import io.github.composegears.valkyrie.generator.jvm.imagevector.common.toResourceText
-import io.github.composegears.valkyrie.parser.svgxml.SvgXmlParser
-import io.github.composegears.valkyrie.parser.svgxml.util.IconType.XML
+import io.github.composegears.valkyrie.parser.unified.ParserType
+import io.github.composegears.valkyrie.parser.unified.SvgXmlParser
+import io.github.composegears.valkyrie.parser.unified.ext.toIOPath
+import io.github.composegears.valkyrie.parser.unified.model.IconType.XML
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
 
@@ -15,8 +17,8 @@ class ImageVectorWithIndentTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `generation without icon pack with indent 1`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/xml/ic_without_path.xml")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/xml/ic_without_path.xml").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
@@ -37,8 +39,8 @@ class ImageVectorWithIndentTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `generation without icon pack with indent 2`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/xml/ic_without_path.xml")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/xml/ic_without_path.xml").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
@@ -59,8 +61,8 @@ class ImageVectorWithIndentTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `generation without icon pack with indent 3`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/xml/ic_without_path.xml")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/xml/ic_without_path.xml").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
@@ -81,8 +83,8 @@ class ImageVectorWithIndentTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `generation without icon pack with indent 6`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/xml/ic_without_path.xml")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/xml/ic_without_path.xml").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
