+++ components/generator/jvm/imagevector/src/test/kotlin/io/github/composegears/valkyrie/generator/jvm/imagevector/ImageVectorWithPreviewTest.kt
@@ -5,8 +5,10 @@ import assertk.assertions.isEqualTo
 import io.github.composegears.valkyrie.extensions.ResourceUtils.getResourcePath
 import io.github.composegears.valkyrie.generator.jvm.imagevector.common.createConfig
 import io.github.composegears.valkyrie.generator.jvm.imagevector.common.toResourceText
-import io.github.composegears.valkyrie.parser.svgxml.SvgXmlParser
-import io.github.composegears.valkyrie.parser.svgxml.util.IconType.XML
+import io.github.composegears.valkyrie.parser.unified.ParserType
+import io.github.composegears.valkyrie.parser.unified.SvgXmlParser
+import io.github.composegears.valkyrie.parser.unified.ext.toIOPath
+import io.github.composegears.valkyrie.parser.unified.model.IconType.XML
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
 
@@ -15,8 +17,8 @@ class ImageVectorWithPreviewTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `androidx preview generation without icon pack`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/xml/ic_without_path.xml")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/xml/ic_without_path.xml").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
@@ -38,8 +40,8 @@ class ImageVectorWithPreviewTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `jetbrains preview generation without icon pack`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/xml/ic_without_path.xml")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/xml/ic_without_path.xml").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
@@ -61,8 +63,8 @@ class ImageVectorWithPreviewTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `androidx preview generation with icon pack`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/xml/ic_without_path.xml")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/xml/ic_without_path.xml").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
@@ -85,8 +87,8 @@ class ImageVectorWithPreviewTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `jetbrains preview generation with icon pack`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/xml/ic_without_path.xml")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/xml/ic_without_path.xml").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
@@ -109,8 +111,8 @@ class ImageVectorWithPreviewTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `androidx preview generation with nested pack`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/xml/ic_without_path.xml")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/xml/ic_without_path.xml").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
@@ -134,8 +136,8 @@ class ImageVectorWithPreviewTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `jetbrains preview generation with nested pack`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/xml/ic_without_path.xml")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/xml/ic_without_path.xml").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
