+++ components/generator/jvm/imagevector/src/test/kotlin/io/github/composegears/valkyrie/generator/jvm/imagevector/SvgWithGradientToImageVectorTest.kt
@@ -5,8 +5,10 @@ import assertk.assertions.isEqualTo
 import io.github.composegears.valkyrie.extensions.ResourceUtils.getResourcePath
 import io.github.composegears.valkyrie.generator.jvm.imagevector.common.createConfig
 import io.github.composegears.valkyrie.generator.jvm.imagevector.common.toResourceText
-import io.github.composegears.valkyrie.parser.svgxml.SvgXmlParser
-import io.github.composegears.valkyrie.parser.svgxml.util.IconType.SVG
+import io.github.composegears.valkyrie.parser.unified.ParserType
+import io.github.composegears.valkyrie.parser.unified.SvgXmlParser
+import io.github.composegears.valkyrie.parser.unified.ext.toIOPath
+import io.github.composegears.valkyrie.parser.unified.model.IconType.SVG
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.EnumSource
 
@@ -15,8 +17,8 @@ class SvgWithGradientToImageVectorTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `svg linear gradient parsing`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/svg/ic_linear_gradient.svg")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/svg/ic_linear_gradient.svg").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
@@ -34,8 +36,8 @@ class SvgWithGradientToImageVectorTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `svg radial gradient parsing`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/svg/ic_radial_gradient.svg")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/svg/ic_radial_gradient.svg").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
@@ -53,8 +55,8 @@ class SvgWithGradientToImageVectorTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `svg linear gradient with stroke parsing`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/svg/ic_linear_gradient_with_stroke.svg")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/svg/ic_linear_gradient_with_stroke.svg").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
@@ -72,8 +74,8 @@ class SvgWithGradientToImageVectorTest {
     @ParameterizedTest
     @EnumSource(value = OutputFormat::class)
     fun `svg with complex gradient`(outputFormat: OutputFormat) {
-        val icon = getResourcePath("imagevector/svg/ic_clip_path_gradient.svg")
-        val parserOutput = SvgXmlParser.toIrImageVector(icon)
+        val icon = getResourcePath("imagevector/svg/ic_clip_path_gradient.svg").toIOPath()
+        val parserOutput = SvgXmlParser.toIrImageVector(parser = ParserType.Jvm, path = icon)
         val output = ImageVectorGenerator.convert(
             vector = parserOutput.irImageVector,
             iconName = parserOutput.iconName,
