+++ components/parser/unified/src/commonMain/kotlin/io/github/composegears/valkyrie/parser/unified/ParserType.kt
@@ -1,0 +1,6 @@
+package io.github.composegears.valkyrie.parser.unified
+
+enum class ParserType {
+    Jvm,
+    Kmp,
+}
