+++ components/parser/unified/src/commonMain/kotlin/io/github/composegears/valkyrie/parser/unified/SvgXmlParser.kt
@@ -1,0 +1,22 @@
+package io.github.composegears.valkyrie.parser.unified
+
+import io.github.composegears.valkyrie.parser.unified.model.IconParserOutput
+import kotlinx.io.files.Path
+
+expect object SvgXmlParser {
+
+    val availableParsers: List<ParserType>
+
+    @Throws(IllegalStateException::class)
+    fun toIrImageVector(
+        parser: ParserType,
+        path: Path,
+    ): IconParserOutput
+
+    @Throws(IllegalStateException::class)
+    fun toIrImageVector(
+        parser: ParserType,
+        value: String,
+        iconName: String,
+    ): IconParserOutput
+}
