+++ components/parser/unified/src/commonMain/kotlin/io/github/composegears/valkyrie/parser/unified/model/IconParserOutput.kt
@@ -1,0 +1,9 @@
+package io.github.composegears.valkyrie.parser.unified.model
+
+import io.github.composegears.valkyrie.ir.IrImageVector
+
+data class IconParserOutput(
+    val iconType: IconType,
+    val irImageVector: IrImageVector,
+    val iconName: String,
+)
