+++ gradle/libs.versions.toml
@@ -7,6 +7,7 @@ kotlin = "2.1.21"
 android-build-tools = "com.android.tools:sdk-common:31.10.0"
 highlights = "dev.snipme:highlights:1.0.0"
 kotlinpoet = "com.squareup:kotlinpoet:2.2.0"
+kotlin-io = "org.jetbrains.kotlinx:kotlinx-io-core:0.7.0"
 leviathan = "io.github.composegears:leviathan:2.0.0"
 leviathan-compose = "io.github.composegears:leviathan-compose:2.0.0"
 tiamat = "io.github.composegears:tiamat:1.5.0"
