+++ settings.gradle.kts
@@ -61,7 +61,7 @@ include("components:parser:jvm:svg")
 include("components:parser:jvm:xml")
 include("components:parser:kmp:xml")
 include("components:parser:kmp:svg")
-include("components:parser:svgxml")
+include("components:parser:unified")
 include("components:psi:iconpack")
 include("components:psi:imagevector")
 include("components:test:coverage")
