+++ tools/idea-plugin/build.gradle.kts
@@ -15,7 +15,7 @@ dependencies {
     implementation(projects.components.generator.jvm.iconpack)
     implementation(projects.components.generator.jvm.imagevector)
     implementation(projects.components.irCompose)
-    implementation(projects.components.parser.svgxml)
+    implementation(projects.components.parser.unified)
     implementation(projects.components.psi.iconpack)
     implementation(projects.components.psi.imagevector)
