+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/mode/iconpack/conversion/IconPackConversionState.kt
@@ -1,7 +1,7 @@
 package io.github.composegears.valkyrie.ui.screen.mode.iconpack.conversion
 
 import io.github.composegears.valkyrie.ir.IrImageVector
-import io.github.composegears.valkyrie.parser.svgxml.util.IconType
+import io.github.composegears.valkyrie.parser.unified.model.IconType
 import io.github.composegears.valkyrie.util.extension.Uuid
 
 sealed interface IconPackConversionState {
