+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/mode/iconpack/newpack/ui/viewmodel/NewPackViewModel.kt
@@ -5,7 +5,7 @@ import io.github.composegears.valkyrie.extensions.safeAs
 import io.github.composegears.valkyrie.generator.jvm.iconpack.IconPackGenerator
 import io.github.composegears.valkyrie.generator.jvm.iconpack.IconPackGeneratorConfig
 import io.github.composegears.valkyrie.generator.jvm.model.IconPack
-import io.github.composegears.valkyrie.parser.svgxml.PackageExtractor
+import io.github.composegears.valkyrie.parser.unified.util.PackageExtractor
 import io.github.composegears.valkyrie.settings.ValkyriesSettings
 import io.github.composegears.valkyrie.ui.di.DI
 import io.github.composegears.valkyrie.ui.extension.updateState
