+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/util/extension/VirtualFile.kt
@@ -1,8 +1,8 @@
 package io.github.composegears.valkyrie.util.extension
 
 import com.intellij.openapi.vfs.VirtualFile
-import io.github.composegears.valkyrie.parser.svgxml.util.isSvgExtension
-import io.github.composegears.valkyrie.parser.svgxml.util.isXmlExtension
+import io.github.composegears.valkyrie.parser.unified.ext.isSvgExtension
+import io.github.composegears.valkyrie.parser.unified.ext.isXmlExtension
 
 inline val VirtualFile.isSvg: Boolean
     get() = extension?.isSvgExtension ?: false
